move folder and fix naming
