+++ src/CSharp/EasyMicroservices.Logger. Log4net/Log4netExtensions.cs
