+++ src/CSharp/EasyMicroservices.Logger. Log4net/Providers/Log4netProvider.cs
