+++ src/CSharp/EasyMicroservices.Logger.Serilog/EasyMicroservices.Logger.Serilog.csproj
@@ -17,7 +17,7 @@
 		<Platforms>AnyCPU;x64;x86</Platforms>
 		<Authors>EasyMicroservices</Authors>
 		<GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-		<Version>0.0.0.2</Version>
+		<Version>0.0.0.3</Version>
 		<Description>logger wrapper</Description>
 		<Copyright>EasyMicroservices@gmail.com</Copyright>
 		<PackageTags>log,logger,logging,serilog</PackageTags>
