+++ src/CSharp/EasyMicroservices.Logger.Tests/BaseProviderAsyncTest.cs
