+++ src/CSharp/EasyMicroservices.Logger.Tests/BaseProviderTest.cs
