+++ src/CSharp/EasyMicroservices.Logger.Tests/EasyMicroservices.Logger.Tests.csproj
