+++ src/CSharp/EasyMicroservices.Logger.Tests/Providers/Log4netProviderTest.cs
