+++ src/CSharp/EasyMicroservices.Logger.Tests/Providers/NLogProviderTest.cs
