+++ src/CSharp/EasyMicroservices.Logger.Tests/Providers/SerilogProviderTest.cs
