+++ src/CSharp/EasyMicroservices.Logger/EasyMicroservices.Logger.csproj
@@ -4,7 +4,7 @@
     <TargetFrameworks>netstandard2.0;netstandard2.1;net6.0</TargetFrameworks>
     <Authors>EasyMicroservices</Authors>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <Version>0.0.0.2</Version>
+    <Version>0.0.0.3</Version>
     <Description>Logger and Auditing system</Description>
     <Copyright>EasyMicroservice@gmail.com</Copyright>
     <PackageTags>logging,auditing</PackageTags>
