+++ src/CSharp/EasyMicroservices.Logger/Interfaces/ILoggerProvider.cs
