+++ src/CSharp/EasyMicroservices.Logger/Interfaces/ILoggerProviderAsync.cs
