+++ src/CSharp/EasyMicroservices.Logger/Models/ExtractedLogInfo.cs
