+++ src/CSharp/EasyMicroservices.Logger/Options/LoggerOption.cs
