+++ src/CSharp/EasyMicroservices.Logger/Providers/BaseProvider.cs
